move from_marbles(), to_iterable(), hot(), parse() from rx.testing.marbles to rx.core.observable.marbles + update tests & examples
