+++ examples/marbles/frommarbles_lookup.py
@@ -2,12 +2,11 @@ import time
 
 import rx
 from rx import concurrency as ccy
-from rx.testing import marbles
 
 lookup0 = {'a': 1, 'b': 3, 'c': 5}
 lookup1 = {'x': 2, 'y': 4, 'z': 6}
-source0 = marbles.from_marbles('a---b----c----|', timespan=0.2, lookup=lookup0)
-source1 = marbles.from_marbles('---x---y---z--|', timespan=0.2, lookup=lookup1)
+source0 = rx.cold('a---b----c----|', timespan=0.2, lookup=lookup0)
+source1 = rx.cold('---x---y---z--|', timespan=0.2, lookup=lookup1)
 
 observable = rx.merge(source0, source1)
